+++ packages/api/src/mapping/DataValueMapping.ts
@@ -46,6 +46,7 @@ export interface DataValueWireToClient {
   }[];
   struct: Record<string, any>;
   set: Set<any>;
+  objectType: string;
 }
 
 /**
@@ -77,6 +78,7 @@ export interface DataValueClientToWire {
     groups: { key: AllowedBucketKeyTypes; value: AllowedBucketTypes }[];
   }[];
   struct: Record<string, any>;
+  objectType: string;
 }
 
 export type AllowedBucketTypes = string | number | boolean;