+++ packages/e2e.generated.catchall/src/generatedNoCheck/index.ts
@@ -1,6 +1,7 @@
 export {
   actionTakesAllParameterTypes,
   assignEmployee1,
+  createFooInterface,
   createMediaObject,
   createMediaViaFunction,
   createOsdkTestObject,