+++ packages/e2e.generated.catchall/src/generatedNoCheck/ontology/actions.ts
@@ -1,5 +1,6 @@
 export { actionTakesAllParameterTypes } from './actions/actionTakesAllParameterTypes.js';
 export { assignEmployee1 } from './actions/assignEmployee1.js';
+export { createFooInterface } from './actions/createFooInterface.js';
 export { createMediaObject } from './actions/createMediaObject.js';
 export { createMediaViaFunction } from './actions/createMediaViaFunction.js';
 export { createOsdkTestObject } from './actions/createOsdkTestObject.js';