+++ packages/e2e.sandbox.todoapp/src/generatedNoCheck2/ontology/actions.ts
@@ -1,2 +1,2 @@
-export { completeTodo } from './actions/completeTodo';
-export { createTodo } from './actions/createTodo';
+export { completeTodo } from './actions/completeTodo.js';
+export { createTodo } from './actions/createTodo.js';