+++ packages/client/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @osdk/client
 
+## 2.1.0-beta.25
+
+### Minor Changes
+
+- c80ca45: Add ability to create interfaces through actions now.
+
+### Patch Changes
+
+- Updated dependencies [c80ca45]
+  - @osdk/generator-converters@2.1.0-beta.25
+  - @osdk/api@2.1.0-beta.25
+  - @osdk/client.unstable@2.1.0-beta.25
+
+## 2.1.0-beta.24
+
+### Minor Changes
+
+- 03f95eb: Add support for interface params in actions.
+- c691c8b: Package now has experimental support for commonjs
+- 3c9ac22: Cleaning up where clause types.
+
+### Patch Changes
+
+- Updated dependencies [03f95eb]
+- Updated dependencies [c691c8b]
+- Updated dependencies [3c9ac22]
+  - @osdk/generator-converters@2.1.0-beta.24
+  - @osdk/api@2.1.0-beta.24
+  - @osdk/shared.client.impl@1.1.0-beta.5
+  - @osdk/shared.net.errors@2.1.0-beta.2
+  - @osdk/shared.net.fetch@1.1.0-beta.3
+  - @osdk/client.unstable@2.1.0-beta.24
+
 ## 2.1.0-beta.23
 
 ### Minor Changes