+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app
 
+## 2.1.0-beta.25
+
+## 2.1.0-beta.24
+
 ## 2.1.0-beta.23
 
 ## 2.1.0-beta.22