+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @osdk/generator
 
+## 2.1.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies [c80ca45]
+  - @osdk/generator-converters@2.1.0-beta.25
+  - @osdk/api@2.1.0-beta.25
+
+## 2.1.0-beta.24
+
+### Minor Changes
+
+- 03f95eb: Add support for interface params in actions.
+- c691c8b: Package now has experimental support for commonjs
+
+### Patch Changes
+
+- Updated dependencies [03f95eb]
+- Updated dependencies [c691c8b]
+- Updated dependencies [3c9ac22]
+  - @osdk/generator-converters@2.1.0-beta.24
+  - @osdk/api@2.1.0-beta.24
+
 ## 2.1.0-beta.23
 
 ### Minor Changes