+++ packages/shared.net.errors/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/shared.net.errors
 
+## 2.1.0-beta.2
+
+### Minor Changes
+
+- c691c8b: Package now has experimental support for commonjs
+
 ## 2.1.0-beta.1
 
 ### Minor Changes